requirements.txt
gradio
python-dotenv
groq
speechrecognition
gtts
torch
torchvision
Pillow
